Create Shoe Fit
Problem
You have three shoes of the same size lying around. Each shoe is either a left shoe (represented using 000) or a right shoe (represented using 111). Given AAA, BBB, CCC, representing the information for each shoe, find out whether you can go out now, wearing one left shoe and one right shoe.
Input Format

    The first line contains an integer TTT, the number of test cases. Then the test cases follow.
    Each test case contains a single line of input, three integers AAA, BBB, CCC.

Output Format

For each test case, output in a single line the answer: 111 if it's possible to go out with a pair of shoes and 000 if not.
Constraints

    1≤T≤81 \leq T \leq 81≤T≤8
    0≤A,B,C≤10 \leq A, B, C \leq 10≤A,B,C≤1

Sample 1:
Input
Output

3
0 0 0
0 1 1
1 0 1

0
1
1

Explanation:

Test Case 111: Since there's no right shoe, it's not possible to go out with a pair of shoes.

Test Case 222: It's possible to go out with a pair of shoes wearing the first and last shoes.

Test Case 333: It's possible to go out with a pair of shoes wearing the first and second shoes.